backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (async)
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Database (sync — for Alembic migrations)
psycopg2-binary==2.9.9

# Redis
redis[asyncio]==5.0.4

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Validation & Config
pydantic==2.7.1
pydantic-settings==2.2.1

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.1.0

# HTTP client (for health check, not production usage)
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-httpx==0.30.0
aiosqlite==0.20.0
